+++ packages/create/src/step-config.ts
@@ -193,11 +193,12 @@ export async function chooseGenConfig(projDir: string, modelPath: string): Promi
 
   // Disable bundled listing and customization features by default
   const bundleListing = 'false'
+  const customConstants = 'false'
   const customLookups = 'false'
   const customOutputs = 'false'
 
   // Add line and write out updated `model.csv`
-  const modelCsvLine = `${initialTime},${finalTime},${datPart},${bundleListing},${customLookups},${customOutputs}`
+  const modelCsvLine = `${initialTime},${finalTime},${datPart},${bundleListing},${customConstants},${customLookups},${customOutputs}`
   const newModelCsvContent = `${modelCsvHeader}\n${modelCsvLine}\n`
   await writeFile(modelCsvFile, newModelCsvContent)