+++ packages/plugin-config/src/processor.ts
@@ -135,6 +135,7 @@ async function processModelConfig(buildContext: BuildContext, options: ConfigPro
     outputs: context.getOrderedOutputs(),
     datFiles: modelOptions.datFiles,
     bundleListing: modelOptions.bundleListing,
+    customConstants: modelOptions.customConstants,
     customLookups: modelOptions.customLookups,
     customOutputs: modelOptions.customOutputs,
     options: options.spec