+++ packages/plugin-wasm/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.2.7](https://github.com/climateinteractive/SDEverywhere/compare/plugin-wasm-v0.2.6...plugin-wasm-v0.2.7) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * peerDependencies
+    * @sdeverywhere/build bumped from ^0.3.7 to ^0.3.10
+
 ## [0.2.6](https://github.com/climateinteractive/SDEverywhere/compare/plugin-wasm-v0.2.5...plugin-wasm-v0.2.6) (2024-12-13)